refactor(ir): change instruction format to fixed word bits (64 bits)

+++ src/main/java/org/elite/ir/IRFunction.java
@@ -32,7 +32,7 @@ public class IRFunction {
     private final int paramCount;
 
     /** Single contiguous code array for all blocks. */
-    private int[] code;
+    private long[] code;
     /** Number of local variable slots */
     private int maxLocals;
     /** Start offset of each basic block in the code array. */
@@ -56,7 +56,7 @@ public class IRFunction {
     }
 
     // Populate IRFunction with code after compilation.
-    void populate(int[] code, int maxLocals, int[] blockOffsets, Object[] constantPool,
+    void populate(long[] code, int maxLocals, int[] blockOffsets, Object[] constantPool,
                   DebugInfo debugInfo, Object[] defaultValues) {
         this.code = code;
         this.maxLocals = maxLocals;
@@ -71,7 +71,7 @@ public class IRFunction {
     public String name()           { return name; }
     public int paramCount()        { return paramCount; }
     public int maxLocals()         { return maxLocals; }
-    public int[] code()            { return code; }
+    public long[] code()            { return code; }
     public int[] blockOffsets()    { return blockOffsets; }
     public Object[] constantPool() { return constantPool; }
     public DebugInfo debugInfo()   { return debugInfo; }
